+++ modules/api/tests/data/ListenerModules.ts
@@ -3,7 +3,7 @@ import type { ModuleListenerInterface } from "@fabernovel/heart-common";
 export const listenerModules: ModuleListenerInterface[] = [
 	{
 		id: "listener1",
-		name: "Heart Listener1",
+		name: "Scodi Listener1",
 		service: {
 			name: "Listener1",
 		},