+++ modules/common/tests/validation/input/analysis/AnalysisInputValidation.test.ts
@@ -135,13 +135,13 @@ describe("Invalid listener options value", () => {
 		{
 			id: "heart-test1",
 			type: "analysis",
-			name: "Heart Test1",
+			name: "Scodi Test1",
 			service: { name: "Test1" },
 		},
 		{
 			id: "heart-test2",
 			type: "analysis",
-			name: "Heart Test2",
+			name: "Scodi Test2",
 			service: { name: "Test2" },
 		},
 	];