+++ modules/mysql/src/client/Client.ts
@@ -50,7 +50,7 @@ export class MySQLClient {
 
 	/**
 	 * Get or create the ORM: avoid the opening of several DB connection at the same or in a very short interval of time.
-	 * As Heart is a CLI tool, having the DB connection opened during a single analysis that longs a couple of minutes is acceptable.
+	 * As Scodi is a CLI tool, having the DB connection opened during a single analysis that longs a couple of minutes is acceptable.
 	 */
 	private async getOrCreateOrm(): Promise<MikroORM<MySqlDriver>> {
 		if (this.#orm === undefined) {